Update scraper.py
I took out the shit calendar trap detector and replaced it with a simple one in the meantime (does not parse any URL with 'calendar' in it) ..  I also added a check that takes out parameters from the URL because this issue caused me to launch back into articles  from 2003.  This code DOES correctly piece together urls that were missing their domain!